api/requirements.txt
# requirements.txt para la API web y descarga periódica
flask
smartsheet-python-sdk
schedule
